+++ R/set.R
@@ -7,12 +7,12 @@
 #' If any query errors result, error messages will be presented to the console.
 #'
 #' @param drill_con drill server connection object setup by \code{drill_connection()}
-#' @param ... named parameters to be sent to ALTER [SYSTEM|SESSION]
+#' @param ... named parameters to be sent to `ALTER SYSTEM` or `ALTER SESSION`
 #' @param type set the \code{session} or \code{system} parameter
 #' @return a \code{tbl} (invisibly) with the \code{ALTER} queries sent and results, including errors.
 #' @references \href{https://drill.apache.org/docs/}{Drill documentation}
 #' @export
-#' @family Dill direct REST API Interface
+#' @family Drill direct REST API Interface
 #' @examples \dontrun{
 #' drill_connection() %>%
 #'   drill_set(exec.errors.verbose=TRUE, store.format="parquet", web.logs.max_lines=20000)
@@ -27,11 +27,11 @@ drill_set <- function(drill_con, ..., type=c("session", "system")) {
 
   purrr::map2(names(params), params, ~sprintf("ALTER %s SET `%s` = %s", type, .x, .y)) %>%
     purrr::map_df(function(x) {
-    y <- drill_query(drill_con, x)
+    y <- drill_query(drill_con, x, .progress=FALSE)
     if (length(y) == 2) {
-      dplyr::data_frame(query=x, param=y$summary, value=y$ok, error_msg=NA)
+      dplyr::tibble(query=x, param=y$summary, value=y$ok, error_msg=NA)
     } else {
-      dplyr::data_frame(query=x, param=NA, value=NA, error_msg=y[[1]])
+      dplyr::tibble(query=x, param=NA, value=NA, error_msg=y[[1]])
     }
   }) -> res
 
@@ -55,7 +55,7 @@ drill_set <- function(drill_con, ..., type=c("session", "system")) {
 #' @param drill_con drill server connection object setup by \code{drill_connection()}
 #' @param ... bare name of system options to reset
 #' @param all if \code{TRUE}, all parameters are reset (\code{...} is ignored)
-#' @family Dill direct REST API Interface
+#' @family Drill direct REST API Interface
 #' @references \href{https://drill.apache.org/docs/}{Drill documentation}
 #' @export
 #' @examples \dontrun{
@@ -70,9 +70,9 @@ drill_system_reset <- function(drill_con, ..., all=FALSE) {
     purrr::map_df(function(x) {
     y <- drill_query(drill_con, x)
     if (length(y) == 2) {
-      dplyr::data_frame(query=x, param=y[[2]]$summary, value=y[[2]]$ok, error=NA)
+      dplyr::tibble(query=x, param=y[[2]]$summary, value=y[[2]]$ok, error=NA)
     } else {
-      dplyr::data_frame(query=x, param=NA, value=NA, error=y[[1]])
+      dplyr::tibble(query=x, param=NA, value=NA, error=y[[1]])
     }
   }) -> res
 
@@ -98,7 +98,7 @@ drill_system_reset <- function(drill_con, ..., all=FALSE) {
 #' @param drill_con drill server connection object setup by \code{drill_connection()}
 #' @param ... bare name of system options to reset
 #' @references \href{https://drill.apache.org/docs/}{Drill documentation}
-#' @family Dill direct REST API Interface
+#' @family Drill direct REST API Interface
 #' @export
 #' @examples \dontrun{
 #' drill_connection() %>% drill_settings_reset(exec.errors.verbose)
@@ -110,9 +110,9 @@ drill_settings_reset <- function(drill_con, ...) {
     purrr::map_df(function(x) {
     y <- drill_query(drill_con, x)
     if (length(y) == 2) {
-      dplyr::data_frame(query=x, param=y[[2]]$summary, value=y[[2]]$ok, error=NA)
+      dplyr::tibble(query=x, param=y[[2]]$summary, value=y[[2]]$ok, error=NA)
     } else {
-      dplyr::data_frame(query=x, param=NA, value=NA, error=y[[1]])
+      dplyr::tibble(query=x, param=NA, value=NA, error=y[[1]])
     }
   }) -> res